requirements.txt
fastapi
uvicorn
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
